backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.22
pymongo==4.16.0
mongomock==4.3.0
httpx==0.28.1
pytest==9.0.2
